+++ 2D_transformation_functions.cpp
@@ -3,8 +3,8 @@
 vector<point2d> translate2d(GLfloat tx, GLfloat ty, vector<point2d> arr) {
 	vector<point2d> ans;
 	for (point2d& it : arr) {
-		GLint nx= it.x + tx;
-		GLint ny= it.y + ty;
+		GLfloat nx= it.x + tx;
+		GLfloat ny= it.y + ty;
 		ans.push_back({nx, ny});
 	}
 	return ans;
@@ -13,8 +13,8 @@ vector<point2d> translate2d(GLfloat tx, GLfloat ty, vector<point2d> arr) {
 vector<point2d> scale2d(GLfloat sx, GLfloat sy, vector<point2d> arr) {
 	vector<point2d> ans;
 	for (point2d& it : arr) {
-		GLint nx = it.x * sx;
-		GLint ny = it.y * sy;
+		GLfloat nx = it.x * sx;
+		GLfloat ny = it.y * sy;
 		ans.push_back({ nx, ny });
 	}
 	return ans;
@@ -24,8 +24,8 @@ vector<point2d> rotate2d(double theta, vector<point2d> arr) {
 	theta = theta*PI/180.0;
 	vector<point2d> ans;
 	for (point2d& it : arr) {
-		GLint nx = it.x * cos(theta) - it.y * sin(theta);
-		GLint ny = it.x * sin(theta) + it.y * cos(theta);
+		GLfloat nx = it.x * cos(theta) - it.y * sin(theta);
+		GLfloat ny = it.x * sin(theta) + it.y * cos(theta);
 		ans.push_back({ nx, ny });
 	}
 	return ans;
@@ -34,8 +34,8 @@ vector<point2d> rotate2d(double theta, vector<point2d> arr) {
 vector<point2d> reflectx2d(vector<point2d> arr) {
 	vector<point2d> ans;
 	for (point2d& it : arr) {
-		GLint nx = it.x;
-		GLint ny = -it.y;
+		GLfloat nx = it.x;
+		GLfloat ny = -it.y;
 		ans.push_back({ nx, ny });
 	}
 	return ans;
@@ -44,8 +44,8 @@ vector<point2d> reflectx2d(vector<point2d> arr) {
 vector<point2d> reflecty2d(vector<point2d> arr) {
 	vector<point2d> ans;
 	for (point2d& it : arr) {
-		GLint nx = -it.x;
-		GLint ny = it.y;
+		GLfloat nx = -it.x;
+		GLfloat ny = it.y;
 		ans.push_back({ nx, ny });
 	}
 	return ans;
@@ -54,8 +54,8 @@ vector<point2d> reflecty2d(vector<point2d> arr) {
 vector<point2d> shear2d(GLfloat shx, GLfloat shy, vector<point2d> arr) {
 	vector<point2d> ans;
 	for (point2d& it : arr) {
-		GLint nx = it.x + shx * it.y;
-		GLint ny = it.y + shy * it.x;
+		GLfloat nx = it.x + shx * it.y;
+		GLfloat ny = it.y + shy * it.x;
 		ans.push_back({ nx, ny });
 	}
 	return ans;
